ai/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
waitress==2.1.2
google-generativeai==0.3.2
requests==2.31.0
gunicorn